quiz/requirements.txt
django
djangorestframework
pymongo
dnspython
Pillow